+++ src/spacegame/Vector2D.java
@@ -1,20 +1,20 @@
 package spacegame;
 
 public class Vector2D {  // Nice
-	public double x;
-	public double y;
+	public float x;
+	public float y;
 	
-	public Vector2D(double x, double y) {
+	public Vector2D(float x, float y) {
 		setVector2D(x,y);
 	}
 	
-	public void setVector2D(double x, double y) {
+	public void setVector2D(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 	
 	public void setVector2D(int x, int y) {
-		this.x = (double) x;
-		this.y = (double) y;
+		this.x = (float) x;
+		this.y = (float) y;
 	}
 }
